+++ utbot-spring-analyzer/src/main/kotlin/org/utbot/spring/config/TestApplicationConfiguration.kt
@@ -3,7 +3,7 @@ package org.utbot.spring.config
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.utbot.spring.post_processors.UtBotBeanFactoryPostProcessor
+import org.utbot.spring.postProcessors.UtBotBeanFactoryPostProcessor
 
 @Configuration
 open class TestApplicationConfiguration {
