+++ utbot-spring-analyzer/src/main/kotlin/org/utbot/spring/postProcessors/UtBotSpringShutdownException.kt
@@ -1,4 +1,4 @@
-package org.utbot.spring.post_processors
+package org.utbot.spring.postProcessors
 
 /**
  * Use this exception to shutdown this application
